+++ packages/ui/src/containers/VerificationCode/use-continue-flow-code-verification.ts
@@ -1,77 +1,47 @@
 import type { EmailVerificationCodePayload, PhoneVerificationCodePayload } from '@logto/schemas';
 import { SignInIdentifier } from '@logto/schemas';
 import { useMemo, useCallback } from 'react';
-import { useTranslation } from 'react-i18next';
-import { useSearchParams, useNavigate } from 'react-router-dom';
+import { useSearchParams } from 'react-router-dom';
 
 import { addProfileWithVerificationCodeIdentifier } from '@/apis/interaction';
 import type { ErrorHandlers } from '@/hooks/use-api';
 import useApi from '@/hooks/use-api';
-import { useConfirmModal } from '@/hooks/use-confirm-modal';
 import useRequiredProfileErrorHandler from '@/hooks/use-required-profile-error-handler';
-import useLinkSocial from '@/hooks/use-social-link-account';
 import type { VerificationCodeIdentifier } from '@/types';
 import { SearchParameters } from '@/types';
-import { formatPhoneNumberWithCountryCallingCode } from '@/utils/country-code';
 
 import useGeneralVerificationCodeErrorHandler from './use-general-verification-code-error-handler';
 import useIdentifierErrorAlert, { IdentifierErrorType } from './use-identifier-error-alert';
+import useLinkSocialConfirmModal from './use-link-social-confirm-modal';
 
 const useContinueFlowCodeVerification = (
   _method: VerificationCodeIdentifier,
   target: string,
   errorCallback?: () => void
 ) => {
-  const { t } = useTranslation();
   const { generalVerificationCodeErrorHandlers, errorMessage, clearErrorMessage } =
     useGeneralVerificationCodeErrorHandler();
   const [searchParameters] = useSearchParams();
-  const { show } = useConfirmModal();
-  const navigate = useNavigate();
-
-  const linkWithSocial = useLinkSocial();
 
   const requiredProfileErrorHandler = useRequiredProfileErrorHandler({ replace: true });
 
   const showIdentifierErrorAlert = useIdentifierErrorAlert();
+  const showLinkSocialConfirmModal = useLinkSocialConfirmModal();
 
   const identifierExistErrorHandler = useCallback(
     async (method: VerificationCodeIdentifier, target: string) => {
       const linkSocial = searchParameters.get(SearchParameters.linkSocial);
 
-      // Should bind with social confirm modal
+      // Show bind with social confirm modal
       if (linkSocial) {
-        const [confirm] = await show({
-          confirmText: 'action.bind_and_continue',
-          cancelText: 'action.change',
-          cancelTextI18nProps: {
-            method: t(
-              `description.${method === SignInIdentifier.Email ? 'email' : 'phone_number'}`
-            ),
-          },
-          ModalContent: t('description.link_account_id_exists', {
-            type: t(`description.${method === SignInIdentifier.Email ? 'email' : 'phone_number'}`),
-            value:
-              method === SignInIdentifier.Email
-                ? target
-                : formatPhoneNumberWithCountryCallingCode(target),
-          }),
-        });
-
-        if (!confirm) {
-          navigate(-1);
-
-          return;
-        }
-
-        await linkWithSocial(linkSocial);
+        await showLinkSocialConfirmModal(method, target, linkSocial);
 
         return;
       }
 
       await showIdentifierErrorAlert(IdentifierErrorType.IdentifierAlreadyExists, method, target);
     },
-    [linkWithSocial, navigate, searchParameters, show, showIdentifierErrorAlert, t]
+    [searchParameters, showIdentifierErrorAlert, showLinkSocialConfirmModal]
   );
 
   const verifyVerificationCodeErrorHandlers: ErrorHandlers = useMemo(
